socketio module adapted
