+++ services/web/server/src/simcore_service_webserver/socketio/config.py
@@ -6,7 +6,7 @@
 from typing import Dict, Optional
 
 import trafaret as T
-from aiohttp import web
+from aiohttp.web import Application
 from pydantic import BaseSettings
 from servicelib.application_keys import APP_CONFIG_KEY
 from socketio import AsyncServer
@@ -26,9 +26,15 @@ class SocketIOSettings(BaseSettings):
     enabled: Optional[bool] = True
 
 
-def get_config(app: web.Application) -> Dict:
+def get_config(app: Application) -> Dict:
     return app[APP_CONFIG_KEY][CONFIG_SECTION_NAME]
 
 
-def get_socket_server(app: web.Application) -> AsyncServer:
+def get_socket_server(app: Application) -> AsyncServer:
     return app[APP_CLIENT_SOCKET_SERVER_KEY]
+
+
+def assert_valid_config(app: Application) -> Dict:
+    cfg = get_config(app)
+    _settings = SocketIOSettings(**cfg)
+    return cfg
